bug 867259: Workaround patch failure on Mac OS X.
By default HFS+ is case insensitive.  This causes a number of netfilter
files within the linux kernel to appear modified on a fresh checkout.  This
in turn causes the patch script to fail due to "existing changes".  Since
B2G has been building fine on Mac with this issue for about year, the thought
is we can safely work around the problem by ignoring the differences between
files with the same names.  Implement this by committing the "changes".